canvi de fitxers, estaven les funcions dels edificis amb les accounts: he fet la separacio
