+++ backend/apps/accounts/serializers.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.db import transaction
 from rest_framework import serializers
 
-from apps.accounts.models import Profile, RoleChoices, Edifici, Localitzacio, Habitatge
+from apps.accounts.models import Profile, RoleChoices
 
 from django.contrib.auth import authenticate
 from rest_framework_simplejwt.tokens import RefreshToken
@@ -112,17 +112,12 @@ class MeSerializer(serializers.ModelSerializer):
         fields = ("id", "email", "first_name", "last_name", "role")
 
 
-# ---------------------------------------------------------------------------
-# Edifici / Habitatge
-# ---------------------------------------------------------------------------
-
 class LocalitzacioResum(serializers.Serializer):
     carrer = serializers.CharField()
     numero = serializers.IntegerField()
     codiPostal = serializers.CharField()
     barri = serializers.CharField()
     zonaClimatica = serializers.CharField()
-
 
 class EdificiResumSerializer(serializers.ModelSerializer):
     localitzacio = LocalitzacioResum(read_only=True)
@@ -131,14 +126,12 @@ class EdificiResumSerializer(serializers.ModelSerializer):
         model = Edifici
         fields = ("idEdifici", "tipologia", "superficieTotal", "puntuacioBase", "localitzacio")
 
-
 class HabitatgeResumSerializer(serializers.ModelSerializer):
     edifici_id = serializers.CharField(source="edifici.idEdifici", read_only=True)
 
     class Meta:
         model = Habitatge
         fields = ("referenciaCadastral", "planta", "porta", "superficie", "edifici_id")
-
 
 # ---------------------------------------------------------------------------
 # Assignació
@@ -162,91 +155,3 @@ class AssignarAdminSerializer(serializers.Serializer):
         if value is not None and not User.objects.filter(pk=value).exists():
             raise serializers.ValidationError("Usuari no trobat.")
         return value
-
-
-class EdificiSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Edifici
-        fields = '__all__'
-
-    # validacio any de construcció
-    def validate_anyConstruccio(self, value):
-        any_actual = date.today().year
-        if value < 1800 or value > any_actual:
-            raise serializers.ValidationError(
-                f"L'any de construcció ha de ser entre 1800 i {any_actual}"
-            )
-        return value
-
-    # validacio superficie
-    def validate_superficieTotal(self, value):
-        if value <= 0:
-            raise serializers.ValidationError(
-                "La superfície total ha de ser més gran que 0."
-            )
-        return value
-
-
-class LocalitzacioSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Localitzacio
-        fields = '__all__'
-
-    # validacio codi postal (5 digits numerics)
-    def validate_codiPostal(self, value):
-        if not re.match(r'^\d{5}$', value):
-            raise serializers.ValidationError(
-                "El format del codi postal és incorrecte. Han de ser 5 dígits."
-            )
-        return value
-
-    # validacio rang latitud (entre -90 i 90)
-    def validate_latitud(self, value):
-        if value < -90 or value > 90:
-            raise serializers.ValidationError(
-                "La latitud ha de ser un valor entre -90 i 90."
-            )
-        return value
-    
-    # validacio rang longitud (entre -180 i 180)
-    def validate_longitud(self, value):
-        if value < -180 or value > 180:
-            raise serializers.ValidationError(
-                "La longitud ha de ser un valor entre -180 i 180."
-            )
-        return value
-
-
-class HabitatgeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Habitatge
-        fields = '__all__'
-
-    # validacio superficie
-    def validate_superficie(self, value):
-        if value <= 0:
-            raise serializers.ValidationError(
-                "La superfície de l'habitatge ha de ser més gran que 0."
-            )
-        return value
-    
-    # validacio any reforma
-    def validate(self, data):
-        any_reforma = data.get('anyReforma')
-        edifici = data.get('edifici')
-
-        if any_reforma is not None:
-            # comprovem que no sigui del futur
-            any_actual = date.today().year
-            if any_reforma > any_actual:
-                raise serializers.ValidationError({
-                    "anyReforma": f"L'any de reforma no pot ser del futur (màxim {any_actual})."
-                })
-            
-            # comprovem que no sigui anterior a la construccio de l'edifici
-            if edifici and any_reforma < edifici.anyConstruccio:
-                raise serializers.ValidationError({
-                    "anyReforma": f"L'any de reforma ({any_reforma}) no pot ser anterior a l'any de construcció de l'edifici ({edifici.anyConstruccio})."
-                })
-
-        return data
