+++ backend/apps/accounts/urls.py
@@ -4,7 +4,6 @@ from rest_framework_simplejwt.views import TokenRefreshView
 from apps.accounts.views import (
     RegisterView, LoginView, LogoutView, MeView,
     MeEdificisView, AssignarResidentView, AssignarAdminEdificiView,
-    EdificiView,
 )
 
 urlpatterns = [
@@ -14,7 +13,6 @@ urlpatterns = [
     path("refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("logout/", LogoutView.as_view(), name="logout"),
     path("me/", MeView.as_view(), name="me"),
-    path("edificis/", EdificiView.as_view(), name="edificis"),
 
     # Consulta: edificis accessibles per l'usuari autenticat
     path("me/edificis/", MeEdificisView.as_view(), name="me-edificis"),
